+++ include/types.hpp
@@ -2,6 +2,6 @@
 #include <vector>
 #include <span>
 
-using t_shape = std::vector<int>;
-using t_indices = std::span<int>;
+using t_shape = std::vector<size_t>;
+using t_indices = std::span<size_t>;
 using t_data = std::vector<float>;
